% Please edit documentation in R/pointdexter.R
\docType{package}
\name{pointdexter-package}
\alias{pointdexter}
\alias{pointdexter-package}
\title{pointdexter: labels longitudinal and latitudinal coordinates located inside a polygon}
\description{
The package labels coordinate pairs located inside a polygon.

For a singular polygon, the label is a logical vector of TRUE and FALSE values. For multiple polygons, the label is a character vector based on the names of each polygon.

The package is designed to work with both \href{https://r-spatial.github.io/sf/}{\code{sf}} and \href{https://www.rdocumentation.org/packages/sp/versions/1.2-5/topics/SpatialPolygonsDataFrame-class}{\code{SpatialPolygonsDataFrame}} objects.
}
\author{
\strong{Maintainer}: Cristian E. Nuno \email{cenuno@syr.edu}
}

